feat: move NotificationProvider into Notifications folder
